+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 5.8.1.0 (beta) - 2025-08-07
 * Fixed issue with layout during exposure template editing.
+* Added twilight offset and dither settings to the exposure template profile table.
 
 ## 5.8.0.0 (beta) - 2025-08-04
 * Added support for exposure rejection for humidity.